adding my logic that was on paper here
It explains my thought process prior to working with the code and then while working with the code. My work changed because there was a realisation of the sorting process that had to be addressed and also the time complexity of the algorithm had to be reduced to improve the problem at hand.